feat: :sparkles: Reactivity
